Created Edit Task Page & Delete Functionality
Created an edit task page, which has a button that can be used to delete the associated task.

Created a TaskForm component to reduce the hard copy of code from AddaskPage and EditTaskPage; though, due to issues, it is unused. If time permits, I will go back and take a look into this.